chore: 개발환경 세팅
- typescript 마이그레이션
- eslint, prettier, husky 세팅
- redux 관련 패키지, axios, styled-components 설치
- react router dom 쓰이는 곳이 없어 삭제
