+++ src/App.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import CommentListContainer from './containers/CommentListContainer';
+import PageListContainer from './containers/PageListContainer';
+import FormContainer from './containers/FormContainer';
+
+function App() {
+  return (
+    <div>
+      <CommentListContainer />
+      <PageListContainer />
+      <FormContainer />
+    </div>
+  );
+}
+
+export default App;
