+++ src/components/Form.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import styled from "styled-components";
+import React from 'react';
+import styled from 'styled-components';
 
 const FormStyle = styled.div`
   & > form {
@@ -11,7 +11,7 @@ const FormStyle = styled.div`
     width: 98%;
     height: 50px;
   }
-  & > form > input[type="text"] {
+  & > form > input[type='text'] {
     padding: 5px 1%;
     width: 98%;
     margin-bottom: 10px;
@@ -37,7 +37,7 @@ function Form() {
         <br />
         <input type="text" name="author" placeholder="작성자" />
         <br />
-        <textarea name="content" placeholder="내용" required></textarea>
+        <textarea name="content" placeholder="내용" required />
         <br />
         <input type="text" name="createdAt" placeholder="2020-05-30" required />
         <br />
