+++ src/containers/CommentListContainer.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import CommentList from "../components/CommentList";
+import React from 'react';
+import CommentList from '../components/CommentList';
 
 function CommentListContainer() {
   return <CommentList />;
